wait before isKeyPressed check
Sometimes pressing a key had no response, because the main event loop
did not wait before checking that a specific key was pressed. Because
of that you had to press a key more than one time or longer than usual
so that the game would catch the event. This is now fixed by calling
the blocking wait_key_pressed procedure.